requirements.txt
pyqt==5.6.2
